basket timer reset for bulk items and regular items - refactored
