+++ client/src/components/organisms/Logout/Logout.js
@@ -3,8 +3,7 @@ import { useCookies } from 'react-cookie';
 import { useHistory } from 'react-router-dom';
 import { AppContext } from '../../../context/app-context';
 import { Button, H2 } from '../../atoms/';
-import { updateItem, deleteItem } from '../../../services/items';
-import { bulkUpdateBatchQuantity } from '../../../utils/bulkUpdateBatchQuantity';
+import { resetBasketItems } from '../../../utils/resetBasketItems';
 
 export const Logout = () => {
   let history = useHistory();
@@ -12,45 +11,10 @@ export const Logout = () => {
   const { setUser, setToken, basket, setBasket, token, setBasketTimer } =
     React.useContext(AppContext);
 
-  const handleLogoutClick = () => {
+  const handleLogoutClick = async () => {
     setUser(null);
-    const batchItemsMap = new Map();
     if (basket && basket.length) {
-      //clear basket from db
-      basket.forEach(async (b) => {
-        // if it's a batch item, we need to delete it and capture the data so we can update the corresponding batch item accordingly
-        // the logic below deals with aggregating all the batch item data into a map, so that only one 'update' method can be called per batch item.
-        if (b.batchId && !b.isTemplateBatchItem) {
-          const sizeType =
-            b.category === 'shoes' ? 'shoeSizes' : 'clothingSizes';
-          const size = b.shoeSize.length > 0 ? b.shoeSize : b.clothingSize;
-          if (!batchItemsMap.has(b.batchId)) {
-            batchItemsMap.set(b.batchId, {});
-          }
-          const batchItemDetails = batchItemsMap.get(b.batchId);
-          if (
-            !Object.prototype.hasOwnProperty.call(batchItemDetails, sizeType)
-          ) {
-            batchItemDetails[sizeType] = {};
-          }
-          const sizeTypeDetails = batchItemDetails[sizeType];
-          if (!Object.prototype.hasOwnProperty.call(sizeTypeDetails, size)) {
-            sizeTypeDetails[size] = 1;
-          } else {
-            sizeTypeDetails[size] += 1;
-          }
-          deleteItem(b._id, token);
-        } else {
-          updateItem(
-            b._id,
-            { inBasket: false, 'statusUpdateDates.inBasketDate': '' },
-            token
-          );
-        }
-      });
-      if (batchItemsMap && batchItemsMap.size > 0) {
-        bulkUpdateBatchQuantity(batchItemsMap, token);
-      }
+      await resetBasketItems(basket, token);
     }
     setBasket(null);
     setBasketTimer(null);
